Added NTN-FR2 FDD bands defined in release 18 in Ka-band range

See merge request oai/openairinterface5g!3414

+++ executables/lte-softmodem.h
@@ -117,7 +117,7 @@ extern uint16_t ue_id_g;
 extern uint16_t node_number;
 
 extern uint64_t downlink_frequency[MAX_NUM_CCs][4];
-extern int32_t  uplink_frequency_offset[MAX_NUM_CCs][4];
+extern int64_t uplink_frequency_offset[MAX_NUM_CCs][4];
 
 extern int rx_input_level_dBm;
 extern openair0_config_t openair0_cfg[MAX_CARDS];
